Add customer_id to accounts
